פרויקט סופי!!!

+++ client/src/Components/Comments.jsx
@@ -17,14 +17,21 @@ export default function Comments(props) {
 
   useEffect(() => {
     async function getComments() {
-      const response = await api.get(`/api/${userId}/comments?postId=${props.postId}`);
+      const response = await api.get(
+        `/api/${userId}/comments?postId=${props.postId}`,
+      );
       setCommentsList(response.data);
+      console.log(commentsList)
     }
     getComments();
   }, [props.postId]);
   async function deleteComment(id) {
-    await api.delete(`/api/${userId}/comments/${id}`);
-    setCommentsList((prev) => prev.filter((comment) => comment.id !== id));
+    if (window.confirm("Are you sure?"));
+    {
+      await api.delete(`/api/${userId}/comments/${id}`);
+     
+      setCommentsList((prev) => prev.filter((comment) => comment.id !== id));
+    }
   }
   async function addNewComment(data) {
     if (data.name.trim() === "") {
@@ -47,8 +54,12 @@ export default function Comments(props) {
   async function updateComment(commentId, updates) {
     const commentToEdit = commentsList.find((p) => p.id === commentId);
     const editedComment = { ...commentToEdit, ...updates };
-    const response = await api.put(`/api/${userId}/comments/${commentId}`, editedComment);
+    const response = await api.put(
+      `/api/${userId}/comments/${commentId}`,
+      editedComment,
+    );
     const updatedComment = response.data;
+    console.log(updatedComment);
     setCommentsList((prev) =>
       prev.map((comment) =>
         comment.id === commentId ? updatedComment : comment,
@@ -89,7 +100,7 @@ export default function Comments(props) {
       {props.showComments ? (
         <>
           <h1>Comments</h1>
-          {commentsList.length > 0 ? (
+          {true ? (
             commentsList.map((comment) => (
               <Comment
                 onDelete={deleteComment}
@@ -99,7 +110,7 @@ export default function Comments(props) {
                 name={comment.name}
                 body={comment.body}
                 email={comment.email}
-                currentUser={comment.email == userEmail}
+                currentUser={comment.userId == userId}
               ></Comment>
             ))
           ) : (
